Create bridge between UIKit and SwiftUI
First, the SwiftUI view is made to conform to
UIViewControllerRepresentable. This has a makeUIViewController() method
which created a PHPickerViewController. A nested Coordinator class
serves as a bridge between the UIKit view controller and the SwiftUI
view. The coordinator has a didFinishPicking method, triggered by iOS
when an image is selected. Finally, the ImagePicker has an @Binding
property so that it can send changes back to a parent view.